intitial approach to oauth-based bitbucket auth
This would probably be the best way forward if Bitbucket support was to be integrated to the upstream.
However, I want to keep the diff small to be able to keep up with the upstream, so I will probably use App Passwords instead. Adding and reverting these work-in-progress changes for future reference.